Extract classes to packages instead
