+++ src/main/java/task/Task.java
@@ -1,0 +1,28 @@
+package task;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " ");
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void markAsUndone() {
+        this.isDone = false;
+    }
+
+    @Override
+    public String toString() {
+        return ("[" + this.getStatusIcon() + "] " + this.description);
+    }
+}
