+++ src/main/java/task/ToDo.java
@@ -1,0 +1,11 @@
+package task;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
